force verification of "iss" and "aud" claims #81
Previously when the following options have been provided

   verify_iss: true, 'iss' => 'acme.org'

and the "iss" claim was missing in the payload - no verification was
executed. The same was true for "aud".

This commit changes the behaviour, so when

   verify_iss: true, 'iss' => 'acme.org'

is set - then the payload is expected to have the "iss" claim and
verification will fail if missing. However, ensure to set `verify_iss`
to `true`, otherwise the verification is not enabled.

This change is backwards compatible, however it might break an app
here or there, because previously the verification did not happen
as expected.

In addition: fixed related test cases and added new ones to ensure
to cover all variants.