Make algorithm option required to verify signature
If the algorithm is not provided and the one in the token's header is used
instead, we might be vulnerable to the attack explained here:

https://auth0.com/blog/critical-vulnerabilities-in-json-web-token-libraries/

For ex:

A server expecting an RSA signed token would do this:

`JWT.decode(token, rsa_public)`

So an attacker can potentially sign a token with HS256 using the same RSA public
key (which is publicly available), and the server will think it's valid.

`JWT.encode({ user: 1 }, rsa_public, 'HS256')`

This doesn't seem to be exploitable right now because the current implementation
of OpenSSL::HMAC.digest expects a string as the key, so if rsa_public is an
OpenSSL::PKey::RSA object, JWT.decode will raise an error. But it would be
better not to depend on this OpenSSL::HMAC.digest behavior