Only issue invalid base64 warning on successful decode
To make the patch minimally invasive, this attempts to decode the
given jwt in strict base64 mode only after a successful full
decode and validation.

Most jwts are small, so this should have minimal performance
impact, though it does require decoding twice.  It is possible to
fix the issue and only decode once, but it would be more invasive.

A big advantage of this approach is the use of :uplevel when
calling Kernel#warn, so that the calling location that triggers
the warning is included in the warning message, greatly simplifying
debugging.

Co-authored-by: Joakim Antman <antmanj@gmail.com>