Backport: Avoid using the same digest across calls (#697) (#701)
Avoid using the same digest across calls (#697)

* Avoid using the same digest across calls

JWT appears to reuse these JWA instances across threads, which
can lead to them stepping on each other via the shared
OpenSSL::Digest instance. This causes decoding to fail
verification, likely because the digest contains an amalgam of
data from the different threads.

This patch creates a new OpenSSL::Digest for each use, avoiding
the threading issue.

Note that the HMAC JWA already calls OpenSSL::HMAC.digest,
avoiding the shared state, and the others do not use digest.

The original code does not fail on CRuby most likely because only
one thread at a time can be calculating a digest against a given
OpenSSL::Digest instance, due to the VM lock.

Fixes jwt/ruby-jwt#696

Addresses the issue reported in jruby/jruby#8504 by @mohamedhafez

* Add #697 to changelog

* Modify Rsa digest name test for new structure

The @digest instance variable now contains the name to the digest
to be used.

See #697

* Add test for concurrent encode/decode using ECDSA

This is adapted from the script in #696 and provides a test for
the ECDSA part of the fix in #697.

* Fixes for Rubocop

Co-authored-by: Charles Oliver Nutter <headius@headius.com>